+++ package.yaml
@@ -1,5 +1,5 @@
 name: PiG
-version: 0.15.0
+version: 0.15.1
 github: PKopel/PiG
 license: BSD3
 author: Paweł Kopel
