+++ src/Import.hs
@@ -3,8 +3,16 @@
 module Import
   ( module RIO
   , module Utils.Util
+  , module Utils.Types
+  , module Utils.Types.App
+  , module Utils.Interp
+  , module Utils.Completion
   )
 where
 
 import           RIO
 import           Utils.Util
+import           Utils.Types
+import           Utils.Types.App
+import           Utils.Completion
+import           Utils.Interp
