+++ src/Utils/Types.hs
@@ -6,7 +6,9 @@
 
 module Utils.Types where
 
-import           Control.Monad.Catch
+import           Control.Monad.Catch            ( MonadMask
+                                                , MonadCatch
+                                                )
 import           Control.Monad.State            ( MonadState
                                                 , StateT(StateT)
                                                 )
@@ -14,32 +16,7 @@ import           Data.List                      ( intercalate
                                                 , (!!)
                                                 )
 import qualified Data.Sequence                 as Seq
-import           Data.Version                   ( Version )
 import           RIO
-import           RIO.Process                    ( HasProcessContext(..)
-                                                , ProcessContext
-                                                )
-import           System.Console.Haskeline       ( Settings )
-
-data Options = Options
-  { optionsVerbose :: !Bool,
-    optionsLoad :: !String
-  }
-
-data App = App
-  { appLogFunc :: !LogFunc,
-    appProcessContext :: !ProcessContext,
-    appOptions :: !Options,
-    appSettings :: !(Settings Interp),
-    appVersion :: !Version
-  }
-
-instance HasLogFunc App where
-  logFuncL = lens appLogFunc (\x y -> x { appLogFunc = y })
-
-instance HasProcessContext App where
-  processContextL =
-    lens appProcessContext (\x y -> x { appProcessContext = y })
 
 class Container c where
   (!?) :: c a -> Int -> Maybe a
